Fix Project

+++ MonitoringAndEvaluationPlatform/Controllers/ReportsController.cs
@@ -295,7 +295,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                         NameAr = m.MinistryDisplayName_AR,
                         ProjectCount = m.Projects.Count,
                         TotalBudget = m.Projects.Sum(p => p.EstimatedBudget),
-                        AmountSpent = ministryProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                        AmountSpent = ministryProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                         IndicatorsPerformance = Math.Round(m.IndicatorsPerformance, 2),
                         DisbursementPerformance = Math.Round(m.DisbursementPerformance, 2)
                     };
@@ -316,7 +316,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                         NameAr = s.AR_Name,
                         ProjectCount = s.Projects.Count,
                         TotalBudget = s.Projects.Sum(p => p.EstimatedBudget),
-                        AmountSpent = sectorProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                        AmountSpent = sectorProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                         IndicatorsPerformance = Math.Round(s.IndicatorsPerformance, 2),
                         DisbursementPerformance = Math.Round(s.DisbursementPerformance, 2)
                     };
@@ -336,7 +336,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                         Name = d.Partner,
                         ProjectCount = d.Projects.Count,
                         TotalBudget = d.Projects.Sum(p => p.EstimatedBudget),
-                        AmountSpent = donorProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                        AmountSpent = donorProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                         IndicatorsPerformance = Math.Round(d.IndicatorsPerformance, 2),
                         DisbursementPerformance = Math.Round(d.DisbursementPerformance, 2)
                     };
@@ -354,7 +354,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                         Name = s.Name,
                         ProjectCount = supervisorProjects.Count,
                         TotalBudget = supervisorProjects.Sum(p => p.EstimatedBudget),
-                        AmountSpent = supervisorProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                        AmountSpent = supervisorProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                         IndicatorsPerformance = supervisorProjects.Any() ? Math.Round(supervisorProjects.Average(p => p.performance), 2) : 0,
                         DisbursementPerformance = supervisorProjects.Any() ? Math.Round(supervisorProjects.Average(p => p.DisbursementPerformance), 2) : 0
                     };
@@ -373,7 +373,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                         Name = pm.Name,
                         ProjectCount = pmProjects.Count,
                         TotalBudget = pmProjects.Sum(p => p.EstimatedBudget),
-                        AmountSpent = pmProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                        AmountSpent = pmProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                         IndicatorsPerformance = pmProjects.Any() ? Math.Round(pmProjects.Average(p => p.performance), 2) : 0,
                         DisbursementPerformance = pmProjects.Any() ? Math.Round(pmProjects.Average(p => p.DisbursementPerformance), 2) : 0
                     };
@@ -401,7 +401,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                         NameAr = g.AR_Name,
                         ProjectCount = allGovProjects.Count,
                         TotalBudget = allGovProjects.Sum(p => p.EstimatedBudget),
-                        AmountSpent = allGovProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                        AmountSpent = allGovProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                         IndicatorsPerformance = Math.Round(allGovProjects.Average(p => p.performance), 2),
                         DisbursementPerformance = Math.Round(allGovProjects.Average(p => p.DisbursementPerformance), 2)
                     };
@@ -419,7 +419,7 @@ namespace MonitoringAndEvaluationPlatform.Controllers
                     NameAr = "الدولة بأكملها",
                     ProjectCount = nationalProjects.Count,
                     TotalBudget = nationalProjects.Sum(p => p.EstimatedBudget),
-                    AmountSpent = nationalProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => plan.Realised) ?? 0),
+                    AmountSpent = nationalProjects.Sum(p => p.Phases?.Where(pp => pp.ActionPlan != null).SelectMany(pp => pp.ActionPlan!.Plans).Sum(plan => (double)plan.Realised) ?? 0),
                     IndicatorsPerformance = Math.Round(nationalProjects.Average(p => p.performance), 2),
                     DisbursementPerformance = Math.Round(nationalProjects.Average(p => p.DisbursementPerformance), 2)
                 });
